backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
qdrant-client==1.7.0
sentence-transformers>=5.1.0
google-generativeai==0.3.2
sendgrid==6.10.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
